When starting a process and invoking the login ui or console, we only check for existance of credentials and not validity. This reduces the amount of roundtrips to the server.
When we try to use the credentials, if it fails, we'll ask to renew.

This way, if the tank command is invoked and it doesn't need to connect to the site, it means that it can run even if credentials are out of date.

is_authenticated has been renamed to has_cached_credentials to reflect this change. Note that in turn, this method is not user anymore.